backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.128.0
Flask==3.1.2
flask-cors==6.0.2
h11==0.16.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pillow==12.1.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
python-nmap==0.7.1
reportlab==4.4.9
requests==2.32.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
Werkzeug==3.1.5
beautifulsoup4
Flask
flask-cors
requests
matplotlib